Module Grpc_test.Config

A configuration is required by the library to know how to run your server and client commands.

The library provides a default configuration that should work for most simple cases. To use it, you have to make sure your server and client commands accept command line parameters specified by the Grpc_discovery library. Then you can use the grpc_discovery function to get a configuration that will work. For more advanced uses, see Advanced API.

type t

module Process_command : sig ... end

val grpc_discovery : run_server_command:Process_command.t -> run_client_command:Process_command.t -> t

If you use Grpc_discovery in your client and server command line, you then only need to supply the path to both commands.

Advanced API

module Client_invocation : sig ... end

module type S = sig ... end

val create : (module S) -> t