Module Grpc_test.Server

type t

A server running that you can connect to during tests.

val listening_on : t -> Eio.Net.Sockaddr.stream

val with_connection : t -> f:(Grpc_client.Connection.t -> Base.unit) -> Base.unit

Initiates a connection to the running server so you can perform RPCs written in OCaml.