Module Grpc_test.Sockaddr_kind

Choosing the kind of sockaddr to use.

Here is what happens in each mode:

Unix_socket

The library creates a temporary file in the file system, and supplies parameters to the server and client command pointing to it, to use it as a unix socket.

Tcp_localhost

The library creates a temporary file in the file system, and supplies parameters to the server and client command pointing to it. This temporary file will serve as a basic service discover via file:

The server command is expected to let the OS choose an available port, listen on the given port and write down that port to the file into a parsable json format. The client command is expected to read the port from the file, and use it to innitiate a connection to this port. The format in used is defined by Grpc_discovery.Discovery_file.

type t =

| Unix_socket

| Tcp_localhost