Module Grpc_test.With_server

type t = {

server : Server.t;

client : ?offline:Base.bool -> Base.string Base.list Base.list -> Base.unit;
(*
client ?offline args is a convenient wrapper for run_client, applied to the given server, and directly introduced to the scope.
*)

}

This is what is introduced to the scope with the with_server function.