Module type Config.S

val run_server_command : listening_config:Grpc_discovery.Listening_config.t -> Process_command.t

This should return a valid command line to run a server that listens to the sockaddr as specified.

val run_client_command : client_invocation:Client_invocation.t -> args:Base.string Base.list -> Process_command.t

Once the server is running, this library will allow running client commands given some arguments. This function specifies the complete command to run. In particular it should add any required parameters in order for the command to connect to the server as specified.